Move unrelated code out of TextIterator https://bugs.webkit.org/show_bug.cgi?id=310147 rdar://172786702
Reviewed by Brent Fulgham and Ryosuke Niwa.

There are a lot of helper functions in TextIterator.cpp that could
be moved into it's own file.

This change also allows other clients to use some of these helper
functions.

* Source/WebCore/Sources.txt:
* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/dom/FragmentDirectiveRangeFinder.cpp:
* Source/WebCore/editing/TextIterator.cpp:
(WebCore::SearchBuffer::SearchBuffer):
(WebCore::SearchBuffer::reachedBreak):
(WebCore::SearchBuffer::search):
(WebCore::foldQuoteMarkAndReplaceNoBreakSpace): Deleted.
(WebCore::foldQuoteMarks): Deleted.
(WebCore::createSearcher): Deleted.
(WebCore::searcher): Deleted.
(WebCore::lockSearcher): Deleted.
(WebCore::unlockSearcher): Deleted.
(WebCore::isKanaLetter): Deleted.
(WebCore::isSmallKanaLetter): Deleted.
(WebCore::composedVoicedSoundMark): Deleted.
(WebCore::isCombiningVoicedSoundMark): Deleted.
(WebCore::containsKanaLetters): Deleted.
(WebCore::normalizeCharacters): Deleted.
(WebCore::isNonLatin1Separator): Deleted.
(WebCore::isSeparator): Deleted.
(WebCore::SearchBuffer::~SearchBuffer): Deleted.
(WebCore::SearchBuffer::isBadMatch const): Deleted.
(WebCore::SearchBuffer::isWordEndMatch const): Deleted.
(WebCore::SearchBuffer::isWordStartMatch const): Deleted.
* Source/WebCore/editing/TextIterator.h:
* Source/WebCore/editing/TextSearchICU.cpp: Added.
(WebCore::createSearcher):
(WebCore::globalSearcher):
(WebCore::ICUSearcher::ICUSearcher):
(WebCore::ICUSearcher::~ICUSearcher):
(WebCore::ICUSearcher::lock):
(WebCore::ICUSearcher::unlock):
(WebCore::ICUSearcher::reset):
(WebCore::ICUSearcher::searcher):
(WebCore::isBadMatch):
(WebCore::isWordStartMatch):
(WebCore::isWordEndMatch):
(WebCore::normalizeCharacters):
(WebCore::foldQuoteMarkAndReplaceNoBreakSpace):
(WebCore::isSeparator):
(WebCore::containsKanaLetters):
(WebCore::foldQuoteMarks):
(WebCore::isKanaLetter):
(WebCore::isSmallKanaLetter):
(WebCore::composedVoicedSoundMark):
(WebCore::isCombiningVoicedSoundMark):
(WebCore::isNonLatin1Separator):
* Source/WebCore/editing/TextSearchICU.h: Added.
* Source/WebCore/page/text-extraction/TextExtraction.cpp:
* Source/WebKit/UIProcess/API/Cocoa/WKWebView.mm:
* Source/WebKit/WebProcess/WebPage/ios/WebPageIOS.mm:

Canonical link: https://commits.webkit.org/309514@main
